Change hostdir variable to clean straneous chars
While using IPv6 hostnames (And maybe hostnames with UTF-8 chars),
2001:0DB8::DEAD:BEEF as example, the expected syntax for hostname is
[2001:0DB8::DEAD:BEEF]. this leads python to create the hostdir as it,
but Java does not understand the path and do not load Linux libraries.

Since dots, colons and any non letter and numeric chars are not
necessary to identify a host, its safe to create a directory with only
letters and numbers to avoid library path issues.